pyproject.toml
[project]
authors = [
  { name = "Konstantin Borovik", email = "kb@lab5.ca" },
]
dependencies = [
  "google-cloud-storage>=2.19.0",
  "logfire[psycopg,system-metrics]>=2.9.0",
  "psycopg[binary,pool]>=3.2.3",
  "pydantic-ai-slim[logfire,openai,vertexai]>=0.0.14",
]
description = "Process invoices using Google Cloud Vision API"
name = "invoice-ocr"
readme = "README.md"
requires-python = ">=3.12"
version = "2024.12.23.post1422"

[project.scripts]
invoice-ocr = "invoice_ocr:main"

[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[dependency-groups]
dev = [
 "python-dotenv>=1.0.1",
 "ruff>=0.8.4",
]

[tool.uv.sources]
invoice_ocr = { workspace = true }

[tool.uv.workspace]
members = [ "invoice_ocr" ]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [ "C90", "F", "N", "N", "PL", "RUF", "SIM", "UP" ]
